Fixed Windows build also for the owncloud storage package
The code for the Windows implementation is taken from fs/local/local_windows.go
